Move tests to their own directory
