+++ src/tests/rgbToCmyk.test.ts
@@ -1,4 +1,4 @@
-import rgbToCmyk from './rgbToCmyk';
+import rgbToCmyk from '../rgbToCmyk';
 import {
   CMYK_BLACK,
   CMYK_BLUE,
@@ -10,7 +10,7 @@ import {
   RGB_GREEN,
   RGB_RED,
   RGB_TEAL,
-} from './testData/colors';
+} from './data/colors';
 
 /**
  * RGB to CMYK
